Fix building issues
