+++ pom.xml
@@ -2,8 +2,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.jetbrains.testPrioritizationPlugin</groupId>
-    <artifactId>testPrioritizationPlugin</artifactId>
+    <groupId>org.jetbrains.teamcity</groupId>
+    <artifactId>test-prioritization-plugin</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
     <properties>
@@ -23,8 +23,8 @@
     </pluginRepositories>
     <modules>
         <module>build</module>
-        <module>testPrioritizationPlugin-agent</module>
-        <module>testPrioritizationPlugin-common</module>
-        <module>testPrioritizationPlugin-server</module>
+        <module>test-prioritization-plugin-agent</module>
+        <module>test-prioritization-plugin-common</module>
+        <module>test-prioritization-plugin-server</module>
     </modules>
 </project>
