+++ test-prioritization-plugin-common/src/main/kotlin/org/jetbrains/teamcity/Constants.kt
@@ -1,0 +1,5 @@
+package org.jetbrains.teamcity
+
+object Constants {
+    const val RUNNER_TYPE = "echoRunnerType"
+}
