+++ test-prioritization-plugin-server/src/main/resources/buildServerResources/viewParams.jsp
@@ -1,0 +1,1 @@
+view
