--- testPrioritizationPlugin-agent/src/main/resources/META-INF/build-agent-plugin-testPrioritizationPlugin.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd"
-       default-autowire="constructor">
-    <bean class="com.jetbrains.testPrioritizationPlugin.EchoFactory"/>
-</beans>
