--- testPrioritizationPlugin-common/src/main/kotlin/com/jetbrains/testPrioritizationPlugin/Constants.kt
@@ -1,6 +1,0 @@
-package com.jetbrains.testPrioritizationPlugin
-
-object Constants {
-    const val RUNNER_TYPE = "echoRunnerType"
-    const val MESSAGE_KEY = "echoMessageKey"
-}
