--- testPrioritizationPlugin-server/src/main/resources/buildServerResources/Hello.jsp
@@ -1,5 +1,0 @@
-<html>
-<body>
-Hi17
-</body>
-</html>
