--- testPrioritizationPlugin-server/src/main/resources/buildServerResources/editParams.jsp
@@ -1,1 +1,0 @@
-edit
