--- testPrioritizationPlugin-server/src/main/resources/buildServerResources/viewParams.jsp
@@ -1,1 +1,0 @@
-view
